bugfixes & tweaks
