+++ pyproject.toml
@@ -38,7 +38,8 @@ license = "python -c 'from scripts.checks import license; license()'"
 set_version = "python -c 'from scripts.set_version import main; main()'"
 
 [tool.poetry.scripts]
-generate_metadata_config = "scripts.ontology.generate_metadata_config:main"
+compute_modules_generate_metadata_config = "compute_modules.bin.ontology.generate_metadata_config:main"
+compute_modules_python_function_infer = "compute_modules.bin.static_inference.infer:main"
 
 [tool.black]
 line_length = 120
